PLZ FIXXXXXXXXXXXXXX TIMERRRRRRRRRRRR

+++ bot.py
@@ -85,12 +85,12 @@ async def textface(ctx, Type):
 async def timer(ctx, timer):
     """Counts down till it's over! Usage: *timer [time in secs]"""
     try:
-        int(timer)
+        float(timer)
     except ValueError:
         await ctx.send("UH OH! Timer did not start. Usage: *timer [time in secs]. Make sure the time is a *whole number*.")
     else:
         await ctx.send("Timer started and rolling! :timer:")
-        await asyncio.sleep(timer)
+        await asyncio.sleep(float(timer))
         await ctx.send("TIME'S UP! :clock:")
